Refactor tests according to file structure
